# Optimize z_stream internal buffer allocations